+++ src/main/java/com/accenture/testingApplication/core/TestingApplication.java
@@ -1,8 +1,11 @@
 package com.accenture.testingApplication.core;
 
+import com.accenture.testingApplication.core.entity.Question;
 import com.accenture.testingApplication.core.logic.InputController;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
 import java.util.Scanner;
 
 
@@ -13,7 +16,10 @@ public class TestingApplication {
 		Scanner in = new Scanner(System.in);
 		SpringApplication.run(TestingApplication.class, args);
 		while (true) {
-			System.out.println(InputController.processingUserInput(in.nextLine()));
+			ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+			InputController inputController = context.getBean("inputController", InputController.class);
+			System.out.println(inputController.processingUserInput(in.nextLine()));
+			context.close();
 		}
 	}
