+++ src/main/java/com/accenture/testingApplication/core/logic/InputController.java
@@ -1,13 +1,14 @@
 package com.accenture.testingApplication.core.logic;
 
-import com.accenture.testingApplication.core.Constant.DialogueConstant;
-import com.accenture.testingApplication.core.connection.ConnectionDataBase;
+import com.accenture.testingApplication.core.сonstant.DialogueConstant;
 import com.accenture.testingApplication.core.entity.User;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class InputController {
+
     private static boolean startModOn;
     private static boolean createQuestionModOn;
     private static boolean openQuestionModOn;
@@ -21,33 +22,35 @@ public class InputController {
     private static boolean userFlow;
     private static boolean administrationFlow;
     private static int flowStep = 0;
-    private static User user = new User();
 
 
-    public static String processingUserInput(String userInput) {
+    public String processingUserInput(String userInput) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+        AdminController adminController = context.getBean("adminController", AdminController.class);
+        UserController userController = context.getBean("userController", UserController.class);
         if (registrationFlow) {
             return registerNewUserProcess(userInput);
         } else if (authorizationFlow) {
             return authorizationUserProcess(userInput);
         } else if (userFlow) {
             if (testingModOn) {
-                return UserController.testPerformer(userInput);
+                return userController.testPerformer(userInput);
             } else {
                 return userProcess(userInput); }
         } else if (administrationFlow) {
             administrationProcess(userInput);
             if (createQuestionModOn) {
-                return (AdminController.createQuestion(userInput));
+                return (adminController.createQuestion(userInput));
             } else if (openQuestionModOn) {
-                return (AdminController.openQuestion(userInput));
+                return (adminController.openQuestion(userInput));
             } else if (deleteQuestionModOn) {
-                return AdminController.deleteQuestion(userInput);
+                return adminController.deleteQuestion(userInput);
             } else if (createTestModOn) {
-                return AdminController.createTest(userInput);
+                return adminController.createTest(userInput);
             } else if (openTestModOn) {
-                return AdminController.openTest(userInput);
+                return adminController.openTest(userInput);
             } else if (deleteTestModOn) {
-                return AdminController.deleteTest(userInput);
+                return adminController.deleteTest(userInput);
             } else if (!administrationFlow) {
                 return DialogueConstant.START_MESSAGE_BOT;
             }
@@ -77,22 +80,28 @@ public class InputController {
         return DialogueConstant.START_MESSAGE_BOT;
     }
 
-    private static String registerNewUserProcess(String currentMessage) {
+    private String registerNewUserProcess(String currentMessage) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+        User user = context.getBean("user", User.class);
         switch (flowStep) {
             case 0:
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_LOGIN_MESSAGE_BOT);
             case 1:
                 user.setLogin(currentMessage);
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_NAME_MESSAGE_BOT);
             case 2:
                 user.setName(currentMessage);
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_PASSWORD_MESSAGE_BOT);
             case 3:
                 user.setPassword(currentMessage);
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_STATUS_MESSAGE_BOT);
             default:
                 if (currentMessage.equals(DialogueConstant.ADMIN_PASSWORD)) {
@@ -100,35 +109,38 @@ public class InputController {
                 } else if (currentMessage.equals(DialogueConstant.NO_ADMIN_MESSAGE)) {
                     user.setAdmin_status(false);
                 } else {
+                    context.close();
                     return (DialogueConstant.MISTAKE_MESSAGE_BOT);
                 }
-                ConnectionDataBase connectionDataBase = new ConnectionDataBase();
                 try {
-                    connectionDataBase.registerUser(user);
+                    user.registerUser(user);
                 } catch (SQLException throwables) {
                     throwables.printStackTrace();
                 }
-                user = new User();
                 flowStopper();
+                context.close();
                 return (DialogueConstant.REGISTER_COMPLETED_MESSAGE_BOT +
                         "\n\n" + DialogueConstant.START_MESSAGE_BOT);
         }
     }
 
-    private static String authorizationUserProcess(String currentMessage) {
+    private String authorizationUserProcess(String currentMessage) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+        User user = context.getBean("user", User.class);
         switch (flowStep) {
             case 0:
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_LOGIN_MESSAGE_BOT);
             case 1:
                 user.setLogin(currentMessage);
                 flowStep++;
+                context.close();
                 return (DialogueConstant.REGISTER_PASSWORD_MESSAGE_BOT);
             default:
                 user.setPassword(currentMessage);
                 user.setAdmin_status(false);
-                ConnectionDataBase connectionDB = new ConnectionDataBase();
-                ResultSet rs = connectionDB.getUser(user);
+                ResultSet rs = user.getUser(user);
                 boolean occurrence = false;
                 try {
                     while (rs.next()) {
@@ -139,7 +151,7 @@ public class InputController {
                 }
                 if (!occurrence) {
                     user.setAdmin_status(true);
-                    rs = connectionDB.getUser(user);
+                    rs = user.getUser(user);
                     boolean adminOccurrence = false;
                     try {
                         while (rs.next()) {
@@ -151,15 +163,18 @@ public class InputController {
                     if (adminOccurrence) {
                         flowStopper();
                         administrationFlow = true;
+                        context.close();
                         return DialogueConstant.AUTHORIZATION_ADMIN_MESSAGE_BOT;
                     } else {
                         flowStopper();
                         modesOff();
+                        context.close();
                         return DialogueConstant.ABSENCE_USER_MESSAGE_BOT;
                     }
                 }
                 flowStopper();
                 userFlow = true;
+                context.close();
                 return DialogueConstant.AUTHORIZATION_USER_MESSAGE_BOT;
         }
     }
