+++ src/main/java/com/accenture/testingApplication/core/logic/UserController.java
@@ -1,12 +1,14 @@
 package com.accenture.testingApplication.core.logic;
 
-import com.accenture.testingApplication.core.Constant.DialogueConstant;
-import com.accenture.testingApplication.core.connection.ConnectionDataBase;
+import com.accenture.testingApplication.core.entity.Question;
+import com.accenture.testingApplication.core.сonstant.DialogueConstant;
 import com.accenture.testingApplication.core.entity.Test;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.util.ArrayList;
 
 public class UserController {
+
     private static int questionNumberInTheTest;
     private static int numberOfCorrectAnswers;
     private static ArrayList questionAnswerParts = new ArrayList();
@@ -14,22 +16,24 @@ public class UserController {
     private static ArrayList testQuestionId = new ArrayList();
     private static int testQuestionIdSize;
 
-    public static String testPerformer(String userMessage) {
+    public String testPerformer(String userMessage) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+        Question question = context.getBean("question", Question.class);
+        Test test = context.getBean("test", Test.class);
         String answer;
-        ConnectionDataBase connectionDB = new ConnectionDataBase();
         if (questionNumberInTheTest == 0) {
-            Test test = new Test(userMessage);
-            for (String question : connectionDB.getTest(test).replace(DialogueConstant.REGULAR_EXPRESSION_QUESTION, "").split(",")) {
-                testQuestionId.add(question);
+            for (String Currentquestion : test.getTest(test).replace(DialogueConstant.REGULAR_EXPRESSION_QUESTION, "").split(",")) {
+                testQuestionId.add(Currentquestion);
                 testQuestionIdSize++;
 
             }
         }
         String[] arrSplit;
         if (questionNumberInTheTest == 0) {
-            questionMessage = connectionDB.getQuestion(testQuestionId.get(questionNumberInTheTest).toString());
+            questionMessage = question.getQuestion(testQuestionId.get(questionNumberInTheTest).toString());
             questionNumberInTheTest++;
             arrSplit = questionMessage.split(DialogueConstant.REGULAR_EXPRESSION_ANSWER);
+            context.close();
             return arrSplit[0];
         } else {
             arrSplit = questionMessage.split(DialogueConstant.REGULAR_EXPRESSION_ANSWER);
@@ -43,13 +47,15 @@ public class UserController {
                 int AnVar = numberOfCorrectAnswers;
                 numberOfCorrectAnswers = 0;
                 testQuestionIdSize = 0;
+                context.close();
                 return DialogueConstant.TEST_FINISH_MESSAGE_BOT +
                         DialogueConstant.NUMBER_CORRECT_ANSWERS +
                         AnVar;
             } else {
-                questionMessage = connectionDB.getQuestion(testQuestionId.get(questionNumberInTheTest).toString());
+                questionMessage = question.getQuestion(testQuestionId.get(questionNumberInTheTest).toString());
                 arrSplit = questionMessage.split(DialogueConstant.REGULAR_EXPRESSION_ANSWER);
                 questionNumberInTheTest++;
+                context.close();
                 return arrSplit[0];
             }
         }
