+++ src/main/resources/applicationContext.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans  xmlns="http://www.springframework.org/schema/beans"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns:context="http://www.springframework.org/schema/context"
+        xsi:schemaLocation="http://www.springframework.org/schema/beans
+        http://www.springframework.org/schema/beans/spring-beans.xsd
+        http://www.springframework.org/schema/context
+        http://www.springframework.org/schema/context/spring-context.xsd">
+
+    <bean id="user"
+          class="com.accenture.testingApplication.core.entity.User">
+    </bean>
+
+    <bean id="question"
+          class="com.accenture.testingApplication.core.entity.Question">
+    </bean>
+
+    <bean id="test"
+          class="com.accenture.testingApplication.core.entity.Test">
+    </bean>
+
+    <bean id="connectionDataBase"
+          class="com.accenture.testingApplication.core.connection.ConnectionDataBase">
+    </bean>
+
+    <bean id="adminController"
+          class="com.accenture.testingApplication.core.logic.AdminController">
+    </bean>
+
+    <bean id="userController"
+          class="com.accenture.testingApplication.core.logic.UserController">
+    </bean>
+
+    <bean id="inputController"
+          class="com.accenture.testingApplication.core.logic.InputController">
+    </bean>
+
+</beans>
